requirements.txt
streamlit
python-dotenv
pinecone
cohere
google-generativeai
tiktoken
pypdf
